Quater4/Class-06-15-SEPTEMBER-2024-langchain-runable-and-agent/pyproject.toml
[tool.poetry]
name = "class-06-15-september-2024-langchain-runable-and-agent"
version = "0.1.0"
description = ""
authors = ["Muhammad Junaid <muhammadjunaid7015140@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.0"
langchain-core = "^0.3.5"
langchain-community = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-google-genai = "^2.0.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
